NAT microservice: handle multiple EdgeGateways
If a vOrg has multiple vDC then it will have multiple EdgeGateways.
We can’t just get the first, as that might be for the wrong vDC.
Therefore find the right one (i.e. the one that has the required
public ips).

Also handle there being multiple Gateway Interfaces of type “uplink”
in an EdgeGateway - find the one with the public ips, rather than 
just picking the last one.